perf: 修改路由守卫
