+++ src/layout/components/NavLeft/index.vue
@@ -52,7 +52,6 @@ export default {
       'sidebar'
     ]),
     activeMenu() {
-      console.log('View routes: ', this.currentViewRoute)
       const route = this.$route
       const { meta, path } = route
       // if set path, the sidebar will highlight the path you set
