+++ src/utils/startup.js
@@ -7,6 +7,7 @@ import 'nprogress/nprogress.css' // progress bar style
 import { getTokenFromCookie } from '@/utils/auth'
 import orgUtil from '@/utils/org'
 import { getCurrentOrg } from '@/api/orgs'
+import { getPropView, getRouteViewRequirePerms, hasPermission } from '@/utils/jms'
 
 const whiteList = ['/login', process.env.VUE_APP_LOGIN_PATH] // no redirect whitelist
 let initial = false
@@ -84,7 +85,11 @@ export async function generatePageRoutes({ to, from, next }) {
     const accessRoutes = await store.dispatch('permission/generateRoutes', { to, from })
 
     // dynamically add accessible routes
+    Vue.$log.debug('All routes: ', accessRoutes)
     router.addRoutes(accessRoutes)
+
+    await store.dispatch('permission/generateViewRoutes', { to, from })
+    console.log('Genepage')
 
     // hack method to ensure that addRoutes is complete
     // set the replace: true, so the navigation will not leave a history record
@@ -104,6 +109,23 @@ export async function checkUserFirstLogin({ to, from, next }) {
   }
 }
 
+export async function changeCurrentViewIfNeed({ to, from, next }) {
+  // const sameView = isSameView(to, from)
+  // const excludes = ['/', '404', '403']
+  // if (excludes.indexOf(to.path) > -1 || sameView) {
+  //   return
+  // }
+  const viewRequirePerms = getRouteViewRequirePerms(to)
+  const hasPerm = hasPermission(viewRequirePerms)
+  Vue.$log.debug('Change current view if need: ', viewRequirePerms, hasPerm)
+  if (hasPerm) {
+    Vue.$log.debug('Has current view perm')
+    return
+  }
+  const view = getPropView()
+  next({ name: view })
+}
+
 export async function startup({ to, from, next }) {
   if (initial) {
     return true
@@ -116,6 +138,7 @@ export async function startup({ to, from, next }) {
   await checkLogin({ to, from, next })
   await changeCurrentOrgIfNeed({ to, from, next })
   await generatePageRoutes({ to, from, next })
+  await changeCurrentViewIfNeed({ to, from, next })
   await checkUserFirstLogin({ to, from, next })
   return true
 }
